# Improve the performance of Basic.__eq__ in Python 2