In PR #12945, `Basic.__eq__` was changed to compare types with `!=` instead of `is
not`. This lead to a performance degradation in Python 2 because of the
`BasicMeta.__cmp__` method (the Basic ordering of classes, see #4269). In Python
2, `!=` calls `__cmp__` if `__ne__` is not defined. This does not affect Python 3
because `__cmp__` is ignored in Python 3.

Removing the ordering of classes has proved to be a more difficult task than I
anticipated (#13059), so for now, we add a workaround. In Python 2, we only
look at `type(self) != type(other)` if the metaclass of `self` actually defines
`__ne__` (bypassing checks to `__cmp__`).

This should be compared against the following (incorrect) patch

```diff
diff --git a/sympy/core/basic.py b/sympy/core/basic.py
index bdfebf24cc..4b3cf77ee7 100644
--- a/sympy/core/basic.py
+++ b/sympy/core/basic.py
@@ -324,7 +324,7 @@ def __eq__(self, other):
             except SympifyError:
                 return NotImplemented

-            if type(self) != type(other):
+            if type(self) is not type(other):
                 return False

         return self._hashable_content() == other._hashable_content()
```

Using the benchmark `ask(Q.real(x**(y/z)), Q.real(x) & Q.real(y/z) &
Q.rational(y/z) & Q.even(z) & Q.positive(x))` (from #12898), in Python 2:

Master (commit before this):

    In [1]: %timeit ask(Q.real(x**(y/z)), Q.real(x) & Q.real(y/z) & Q.rational(y/z) & Q.even(z) & Q.positive(x))
    1 loop, best of 3: 4.57 s per loop

The above (incorrect) patch:

    In [1]: %timeit ask(Q.real(x**(y/z)), Q.real(x) & Q.real(y/z) & Q.rational(y/z) & Q.even(z) & Q.positive(x))
    1 loop, best of 3: 3.76 s per loop

This commit:

    In [1]: %timeit ask(Q.real(x**(y/z)), Q.real(x) & Q.real(y/z) & Q.rational(y/z) & Q.even(z) & Q.positive(x))
    1 loop, best of 3: 3.85 s per loop

So it isn't quite as fast, but it's about as fast as we can get while
maintaining the proper metaclass `__eq__` checks (which are needed for the
`Function('f', real=True)` feature to work).

Also worth noting that Python 3 master is faster still:

    In [1]: %timeit ask(Q.real(x**(y/z)), Q.real(x) & Q.real(y/z) & Q.rational(y/z) & Q.even(z) & Q.positive(x))
    3.07 s ± 72.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

Additionally, it should be pointed out that benchmarks that make direct use of
undefined functions may still be slower than before #12945, due to the more
advanced logic that was added there (I haven't looked into whether
`UndefinedFunction.__eq__` itself could be sped up).

Fixes https://github.com/sympy/sympy/issues/14243