@asmeurer 2018-05-17
Ping @moorepants can you run asv compare against master and this branch in Python 2? I never seem to get any kind of consistent results from asv (with anything, not just this benchmark). I'm honestly starting to have serious doubts about it. Maybe my computer isn't configured properly for it, I don't know. 

The above test seems to show that this has a speedup in Python 2, but I'd like to see just how much this affects the other benchmarks that I know were slowed down by #12945.