@asmeurer 2018-06-28
I never got a benchmarking update. I hope this is enough. Merging because this is blocking the release.